Refactors in order to establish a consistent directory structure.
